+++ main.py
@@ -1,17 +1,16 @@
-import sys, windows, load
+import sys, windows
 from PyQt6 import QtWidgets
 
 if __name__ == "__main__":
     # create app
+    print("Starting app")
     app = QtWidgets.QApplication(sys.argv)
     mainWindow = QtWidgets.QMainWindow()
-    windowManager = windows.WindowManager(mainWindow)
 
-    # load and insert profiles
-    profiles = load.createProfileList()
-    for profile in profiles:
-        windowManager.addProfile(profile)
+    # create UI and load from data from json file
+    windowManager = windows.WindowManager(mainWindow, processEventsFunc=lambda : app.processEvents())
 
-    # show window and configure exit button
+    # show window and close temporary loading window
+    print("Done! Displaying window")
     mainWindow.showMaximized()
     sys.exit(app.exec())
