--- mods123.json
@@ -1,76 +1,0 @@
-[
-    {
-        "priorityname": "High Priority",
-        "priorityr": 255,
-        "priorityg": 85,
-        "priorityb": 0,
-        "name": "Lithium",
-        "id": "gvQqBUqZ",
-        "url": "https://modrinth.com/mod/lithium",
-        "versions": [
-            "1.16.2",
-            "1.16.3",
-            "1.16.4",
-            "1.16.5",
-            "1.17",
-            "1.17.1",
-            "1.18",
-            "1.18.1",
-            "1.18.2",
-            "1.19",
-            "1.19.1",
-            "1.19.2",
-            "1.19.3",
-            "1.19.4",
-            "1.20",
-            "1.20.1",
-            "1.20.2",
-            "1.20.3",
-            "1.20.4",
-            "1.20.5",
-            "1.20.6",
-            "1.21",
-            "1.21.1",
-            "1.21.2",
-            "1.21.3",
-            "1.21.4",
-            "1.21.5"
-        ]
-    },
-    {
-        "priorityname": "High Priority",
-        "priorityr": 255,
-        "priorityg": 85,
-        "priorityb": 0,
-        "name": "Sodium",
-        "id": "AANobbMI",
-        "url": "https://modrinth.com/mod/sodium",
-        "versions": [
-            "1.16.3",
-            "1.16.4",
-            "1.16.5",
-            "1.17.1",
-            "1.18",
-            "1.18.1",
-            "1.18.2",
-            "1.19",
-            "1.19.1",
-            "1.19.2",
-            "1.19.3",
-            "1.19.4",
-            "1.20",
-            "1.20.1",
-            "1.20.2",
-            "1.20.3",
-            "1.20.4",
-            "1.20.5",
-            "1.20.6",
-            "1.21",
-            "1.21.1",
-            "1.21.2",
-            "1.21.3",
-            "1.21.4",
-            "1.21.5"
-        ]
-    }
-]
